started work on sffinfo command.  fixed bug across all paralellized commands if the file position was too big to fit in an int.

+++ chimera.h
@@ -77,9 +77,9 @@ struct sim {
 };
 
 struct linePair {
-			int start;
-			int end;
-			linePair(int i, int j) : start(i), end(j) {}
+			unsigned long int start;
+			unsigned long int end;
+			linePair(unsigned long int i, unsigned long int j) : start(i), end(j) {}
 			linePair(){}
 };
